Correction for the OpenCL program, specifically for the "inverse x" forces.
Complete overhaul of the "OpenCL" class due to malfunction on systems
with 2+ openCL devices. Architecture changed to one context per device.
One context per multiple devices was causing weird behaviour because
multiple devices tried to write in the same buffer at the same time as
it is shared across devices in the context. The new implementation also
allowed devices of diffenrent platforms to be used.